+++ src/main/java/dev/doctor4t/ratatouille/client/util/CustomModelArmorUtil.java
@@ -6,9 +6,7 @@ import net.minecraft.client.model.Dilation;
 import net.minecraft.client.model.TexturedModelData;
 import net.minecraft.client.render.entity.EntityRendererFactory;
 import net.minecraft.client.render.entity.model.EntityModelLayer;
-import net.minecraft.entity.EquipmentSlot;
 import net.minecraft.entity.LivingEntity;
-import net.minecraft.item.Item;
 import net.minecraft.util.Identifier;
 import org.jetbrains.annotations.NotNull;
 
@@ -17,13 +15,13 @@ import java.util.Map;
 import java.util.function.Function;
 
 public class CustomModelArmorUtil {
-    public static final Map<SetItems, SetRenderData> CUSTOM_ARMOR_MODELS = new HashMap<>();
+    public static final Map<ArmorDisplayConditions, SetRenderData> CUSTOM_ARMOR_MODELS = new HashMap<>();
 
     public static @NotNull EntityModelLayer registerModelLayerMain(Identifier identifier) {
         return new EntityModelLayer(identifier, "main");
     }
 
-    public static void registerCustomArmor(Identifier id, SetItems setItems, CustomArmorModelDefinition armorModelDefinition, int textureWidth, int textureHeight) {
+    public static void registerCustomArmor(Identifier id, ArmorDisplayConditions displayConditions, CustomArmorModelDefinition armorModelDefinition, int textureWidth, int textureHeight) {
         EntityModelLayer modelLayer = registerModelLayerMain(id);
 
         TexturedModelData modelData = TexturedModelData.of(CustomBipedArmorModel.getModelData(armorModelDefinition::addModelParts, Dilation.NONE), textureWidth, textureHeight);
@@ -34,32 +32,9 @@ public class CustomModelArmorUtil {
                 context -> new CustomBipedArmorModel<>(context.getPart(modelLayer), armorModelDefinition)
         );
 
-        CUSTOM_ARMOR_MODELS.put(setItems, setRenderData);
+        CUSTOM_ARMOR_MODELS.put(displayConditions, setRenderData);
 
         EntityModelLayerRegistry.registerModelLayer(setRenderData.modelLayer(), setRenderData::modelData);
-    }
-
-    public record SetItems(
-            Item helmetItem,
-            Item chesplateItem,
-            Item leggingsItem,
-            Item bootsItem) {
-
-        public boolean shouldDisplayHelmet(LivingEntity livingEntity) {
-            return livingEntity.getEquippedStack(EquipmentSlot.HEAD).isOf(this.helmetItem);
-        }
-
-        public boolean shouldDisplayChestplate(LivingEntity livingEntity) {
-            return livingEntity.getEquippedStack(EquipmentSlot.CHEST).isOf(this.chesplateItem);
-        }
-
-        public boolean shouldDisplayLeggings(LivingEntity livingEntity) {
-            return livingEntity.getEquippedStack(EquipmentSlot.LEGS).isOf(this.leggingsItem);
-        }
-
-        public boolean shouldDisplayBoots(LivingEntity livingEntity) {
-            return livingEntity.getEquippedStack(EquipmentSlot.FEET).isOf(this.bootsItem);
-        }
     }
 
     public record SetRenderData(